Moved some files around to work correctly with gunicorn
