+++ app/templates/background.js
