+++ app/templates/server.js
