+++ wsgi.py
@@ -2,7 +2,7 @@
 This file describes how gunicron will route traffic    
 """
 
-from .app import app
+from app.app import app
 
 if __name__ == "__main__":
         app.run()
